+++ sae_lens/__init__.py
@@ -1,15 +1,15 @@
 __version__ = "2.1.3"
 
+from .cache_activations_runner import CacheActivationsRunner
+from .pretokenize_runner import pretokenize_runner
+from .sae_training_runner import SAETrainingRunner
 from .training.activations_store import ActivationsStore
-from .training.cache_activations_runner import CacheActivationsRunner
 from .training.config import (
     CacheActivationsRunnerConfig,
     LanguageModelSAERunnerConfig,
     PretokenizeRunnerConfig,
 )
 from .training.evals import run_evals
-from .training.lm_runner import SAETrainingRunner
-from .training.pretokenize_runner import pretokenize_runner
 from .training.session_loader import LMSparseAutoencoderSessionloader
 from .training.sparse_autoencoder import SparseAutoencoder
