+++ tutorials/mamba_train_example.py
@@ -6,10 +6,11 @@ import torch
 
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), ".."))
 
+from sae_lens.sae_training_runner import SAETrainingRunner
+
 # run this as python3 tutorials/mamba_train_example.py
 # i.e. from the root directory
 from sae_lens.training.config import LanguageModelSAERunnerConfig
-from sae_lens.training.lm_runner import SAETrainingRunner
 
 cfg = LanguageModelSAERunnerConfig(
     # Data Generating Function (Model + Training Distibuion)
